Added ‘Marathon’ based helper to call setup script
It is a special case here — we wrote setup script that depends on the repo itself, we had to put such script in a subfolder and implemente it as another package using SwiftPM. In other cases ‘XCERepoConfigurator’ is supposed to be used as dependncy for ‘Marathon’ scripts or other SPM packages. So we add this helper ‘Marathon’ script for seamless experience.